refactor: update imports to reflect controller package move
